Add Activity Detail Tips and change some layout
