+++ app/src/main/res/layout/activity_detail.xml
@@ -47,15 +47,48 @@
                     android:layout_centerInParent="true"/>
             </RelativeLayout>
 
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/item_title"
+                android:layout_marginTop="8dp"
+                android:textSize="24sp"
+                android:textStyle="bold"
+                android:layout_marginHorizontal="14dp"
+                />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/rp_1000"
+                android:layout_marginHorizontal="14dp"
+                android:layout_marginTop="4dp"
+                android:textSize="18sp"
+                android:textColor="@color/light_green"
+                />
+
+            <com.google.android.material.button.MaterialButton
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/contact_seller"
+                android:textColor="@color/light_green"
+                android:layout_marginVertical="8dp"
+                android:layout_marginHorizontal="14dp"
+                android:backgroundTint="@color/white"
+                app:strokeColor="@color/light_green"
+                app:strokeWidth="1dp"
+                />
+
             <LinearLayout
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:orientation="horizontal"
                 tools:ignore="UseCompoundDrawables">
+
                 <ImageView
                     android:id="@+id/img_user_uploader"
-                    android:layout_width="50dp"
-                    android:layout_height="50dp"
+                    android:layout_width="40dp"
+                    android:layout_height="40dp"
                     android:layout_marginVertical="10dp"
                     android:layout_marginStart="10dp"
                     android:src="@drawable/ic_baseline_account_circle_24"
@@ -83,84 +116,7 @@
                 android:textSize="18sp"
                 />
 
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="32dp"
-                android:layout_marginTop="32dp"
-                android:background="@color/light_green"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="14dp"
-                    android:text="@string/comment"
-                    android:textColor="@color/white"
-                    android:textSize="18sp"
-                    android:textStyle="bold"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" />
-            </androidx.constraintlayout.widget.ConstraintLayout>
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/rv_comment"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                />
         </LinearLayout>
     </androidx.core.widget.NestedScrollView>
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:cardBackgroundColor="@color/white"
-        app:contentPaddingLeft="18dp"
-        app:contentPaddingRight="18dp"
-        app:contentPaddingTop="10dp"
-        app:contentPaddingBottom="10dp"
-        app:cardElevation="14dp"
-        android:layout_alignParentBottom="true">
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content">
-            <ImageView
-                android:id="@+id/img_user_owner"
-                android:layout_width="50dp"
-                android:layout_height="50dp"
-                android:src="@drawable/ic_baseline_account_circle_24"
-                android:contentDescription="@string/user_profile_picture"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toStartOf="@+id/edt_comment"/>
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/edt_comment"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="4dp"
-                android:layout_marginEnd="4dp"
-                android:hint="@string/add_a_comment"
-                android:minHeight="48dp"
-                android:textSize="14sp"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toEndOf="@+id/img_user_owner"
-                app:layout_constraintEnd_toStartOf="@id/btn_send"/>
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btn_send"
-                android:layout_width="32dp"
-                android:layout_height="32dp"
-                android:background="@drawable/ic_baseline_send_24"
-                android:backgroundTint="@color/green"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                />
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.cardview.widget.CardView>
 
 </RelativeLayout>
